requirements.txt
chromadb
fastapi
langchain
langchain_openai
langchain_community
nltk
openai
psycopg2
pydantic
python-dotenv
python-multipart
trulens_eval==0.21.0
uvicorn